Refactored deleting tracking Fixed date issues, refactoring

+++ banners/data/delete/deleting_repository.py
@@ -1,10 +1,12 @@
 import json
 import time
+from typing import Dict
 
 from flask import Response
 
 from application import app_sqlite_db
 from application.base_response import BaseResponse
+from banners.data.actions.action_item import AdminActionModel, AdminAction
 from banners.data.admin_repository import admin_validation
 from banners.data.firebase.firestore_repository import get_shared_banner_ref
 from banners.types.deleted_banner import DeletedBannerModel, DeletedBannerModelEncoder
@@ -12,29 +14,20 @@ from cat.utils.telegram_utils import send_telegram_msg_to_me
 from config import BE_PAGE_SIZE
 
 
-def get_deleted_banners(page=0) -> []:
-    deleted_banners = app_sqlite_db.session.query(DeletedBannerModel).order_by(DeletedBannerModel.date.desc())
-
-    selection = []
-
-    last_item_index = (page + 1) * BE_PAGE_SIZE
-    first_item_index = page * BE_PAGE_SIZE
-
-    if deleted_banners.count() < last_item_index:
-        last_item_index = deleted_banners.count()
-
-    for banner in deleted_banners[first_item_index:last_item_index]:
-        selection.append(banner)
-
-    return selection
+def get_deleted_banners(page=0):
+    return (
+        app_sqlite_db.session.query(DeletedBannerModel)
+        .order_by(DeletedBannerModel.date.desc())
+        .offset(page * BE_PAGE_SIZE)
+        .limit(BE_PAGE_SIZE)
+        .all()
+    )
 
 
-def delete_banner(request) -> Response:
+def delete_banner(request, mock: bool = False) -> Response:
     content = request.args.to_dict()
 
-    check_result = admin_validation(content)
-
-    if not check_result.success:
+    if not (check_result := admin_validation(content)).success:
         return check_result.to_response()
 
     admin_id = content["admin"]
@@ -46,42 +39,28 @@ def delete_banner(request) -> Response:
     if banner_data is None:
         return BaseResponse(False, f"Banner with id {banner_id} not found!", content).to_response()
 
-    send_telegram_msg_to_me(f"Admin with id {admin_id} requested deletion of this banner {banner_id}\n\n{banner_data}")
+    store_deleted_banner(banner_id, admin_id, banner_data)
 
-    app_sqlite_db.session.add(DeletedBannerModel(id=banner_id, admin_id=admin_id,
-                                                 layers=str(json.dumps(banner_data['mlayers'], ensure_ascii=False, cls=DeletedBannerModelEncoder)).replace("\'", "\""),
-                                                 content=json.dumps(banner_data), date=time.time()))
-    app_sqlite_db.session.commit()
-
-    try:
-        banner_ref.delete()
-    except Exception as error:
-        return BaseResponse(False, str(error), content).to_response()
+    if not mock:
+        try:
+            banner_ref.delete()
+        except Exception as error:
+            return BaseResponse(False, str(error), content).to_response()
 
     return BaseResponse(True).to_response()
 
-def test_deleted_banner(request) -> Response:
-    content = request.args.to_dict()
 
-    check_result = admin_validation(content)
+def store_deleted_banner(banner_id: str, admin_id: str, banner_data: Dict):
+    app_sqlite_db.session.add(DeletedBannerModel.build(
+        banner_id=banner_id,
+        admin_id=admin_id,
+        banner_data=banner_data
+    ))
 
-    if not check_result.success:
-        return check_result.to_response()
+    app_sqlite_db.session.add(AdminActionModel.build(
+        admin_id=admin_id,
+        action_info=banner_data,
+        action=AdminAction.Deleted
+    ))
 
-    admin_id = content["admin"]
-    banner_id = content["id"]
-
-    banner_ref = get_shared_banner_ref(banner_id)
-    banner_data = banner_ref.get().to_dict()
-
-    if banner_data is None:
-        return BaseResponse(False, f"Banner with id {banner_id} not found!", content).to_response()
-
-    send_telegram_msg_to_me(f"Admin with id {admin_id} requested deletion of this banner {banner_id}\n\n{banner_data}")
-
-    app_sqlite_db.session.add(DeletedBannerModel(id=banner_id, admin_id=admin_id,
-                                                 layers=str(json.dumps(banner_data['mlayers'], ensure_ascii=False, cls=DeletedBannerModelEncoder)).replace("\'", "\""),
-                                                 content=json.dumps(banner_data), date=time.time()))
     app_sqlite_db.session.commit()
-
-    return BaseResponse(True).to_response()
